fix: use float16 with masked_scatter_ patch to avoid OOM and dtype mismatch
Float32 OOMs on a single T4 (15GB). Switch back to float16 (~3GB) and
monkey-patch masked_scatter_ to auto-cast source dtype, fixing the
bfloat16 autocast issue on T4 without needing float32.

+++ src/config.py
@@ -11,7 +11,7 @@ class ModelConfig:
     """Configuration for the DeepSeek OCR model."""
 
     model_id: str = "deepseek-ai/DeepSeek-OCR-2"
-    torch_dtype: str = "float32"
+    torch_dtype: str = "float16"
     base_size: int = 1024
     image_size: int = 768
     crop_mode: bool = True
